Styled home page and fixed modals

+++ react-app/src/components/OneDish/index.js
@@ -1,6 +1,4 @@
 import {useDispatch, useSelector} from 'react-redux'
-// import { getReviewsThunk } from '../../store/review'
-// import CreateReviewForm from '../CreateReviewForm/CreateReviewForm'
 import {useEffect, useState} from 'react'
 import {useParams} from 'react-router-dom'
 import { deletePrimaryReviewThunk, getOnePrimaryReviewThunk } from '../../store/primaryReview'
@@ -9,9 +7,12 @@ import EditDish from '../EditDish'
 import { NavLink } from 'react-router-dom'
 import EditSecondaryDish from '../EditSecondaryDish'
 import { deleteSecondaryReviewThunk } from '../../store/secondaryReview'
+import { Modal } from '../../context/modal'
+import './index.css'
 
 export default function OneDish () {
-
+  const [showModal, setShowModal] = useState(false);
+  const [showModal2, setShowModal2] = useState(false);
   let {dishId} = useParams()
   dishId = parseInt(dishId)
   const dispatch = useDispatch()
@@ -37,38 +38,56 @@ export default function OneDish () {
   }
 
 return (
-<div>
-      <div>
-      {primaryDish?.name}
-      {primaryDish?.description}
-      <img className="smallImgs" src={primaryDish?.image}></img>
-
-      {primaryDish?.address}
-      {primaryDish?.rating}
-      </div>
-      <EditDish/>
-      {/* {userId === primaryDish.id && */}
-        <div>
-        {/* <NavLink to={`/`}> */}
-        <button className='crudButton' onClick={()=> {dispatch(deletePrimaryReviewThunk(dishId))}}>Delete Review</button>
-        {/* </NavLink> */}
-        </div>
-        {/* // } */}
-        {/* <div className='allReviews'> */}
   <div>
-    <h2>Additional Reviews</h2>
-    {secondaryDishes.map(secondaryDish =>
-    <div>
-      <div>{secondaryDish.name}</div>
-      <img src={secondaryDish.image}></img>
-      <div>{secondaryDish.address}</div>
-      <div>{secondaryDish.description}</div>
-      <EditSecondaryDish/>
-      <button className='crudButton' onClick={()=> {dispatch(deleteSecondaryReviewThunk(secondaryDish.id))}}>Delete Review</button>
+    <div className='primeDish'>
+      <div className='primeImage'>
+        <img src={primaryDish?.image}></img>
+      </div>
+      <div className='primeInfoAndButtons'>
+        <div className='primeInfo'>
+          <h1>The Winning Dish</h1>
+        <div>Dish: {primaryDish?.name}</div>
+        <div>Description: {primaryDish?.description}</div>
+        <div>Address: {primaryDish?.address}</div>
+        <div>Rating: {primaryDish?.rating}</div>
+        </div>
+      <button onClick={() => setShowModal(true)}>Edit your Dish</button>
+      {showModal && (
+        <Modal onClose={() => setShowModal(false)}>
+          <EditDish setShowModal={setShowModal} />
+        </Modal>
+      )}
+        <NavLink to={`/`}>
+          <button className='crudButton' onClick={() => { dispatch(deletePrimaryReviewThunk(dishId)) }}>Delete Review</button>
+        </NavLink>
+      </div>
     </div>
-    )}
+    <h2 className='addReview'>Additional Reviews</h2>
+    <div className='additionalReviews'>
+      {secondaryDishes.map(secondaryDish =>
+        <div className='secondaryDish'>
+          <div className='secondDishDiv'><img className='secondDishImg' src={secondaryDish.image}></img></div>
+          <div>
+            <div className='secondReviewText'>
+              <div>{secondaryDish.name}</div>
+              <div>{secondaryDish.address}</div>
+              <div>{secondaryDish.description}</div>
+            </div>
+            <div className='editAndDeleteSecond'>
+              <button onClick={() => setShowModal2(true)}>Edit your Dish</button>
+              {showModal2 && (
+              <Modal onClose={() => setShowModal2(false)}>
+                <EditSecondaryDish setShowModal2={setShowModal2} />
+              </Modal>
+              )}
+              <button className='crudButton' onClick={() => { dispatch(deleteSecondaryReviewThunk(secondaryDish.id)) }}>Delete Review</button>
+              </div>
+          </div>
+        </div>
+      )}
+    </div>
+
   </div>
-</div>
 )
 
 }
